requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
blinker==1.9.0
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cymem==2.0.8
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
executing==2.1.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
langcodes==3.5.0
langgraph==0.2.53
langgraph-checkpoint==2.0.7
langgraph-sdk==0.1.40
langsmith==0.1.143
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.14.2
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
openai==1.55.0
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.0.0
pydantic==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.27.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
stack-data==0.6.3
streamlit==1.40.1
tenacity==9.0.0
thinc==8.2.5
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.3
typer==0.13.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
wheel==0.45.0
wrapt==1.16.0
yarl==1.17.2
